+++ Forge/src/main/java/fuzs/puzzleslib/network/ForgeNetworkHandler.java
@@ -2,15 +2,10 @@ package fuzs.puzzleslib.network;
 
 import fuzs.puzzleslib.core.DistTypeConverter;
 import fuzs.puzzleslib.proxy.Proxy;
-import net.minecraft.client.Minecraft;
-import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
 import net.minecraft.network.protocol.Packet;
-import net.minecraft.resources.ResourceKey;
 import net.minecraft.resources.ResourceLocation;
-import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.world.entity.player.Player;
-import net.minecraft.world.level.Level;
 import net.minecraftforge.common.util.LogicalSidedProvider;
 import net.minecraftforge.fml.LogicalSide;
 import net.minecraftforge.network.NetworkDirection;
@@ -18,8 +13,6 @@ import net.minecraftforge.network.NetworkEvent;
 import net.minecraftforge.network.NetworkRegistry;
 import net.minecraftforge.network.simple.SimpleChannel;
 
-import javax.annotation.Nullable;
-import java.util.Objects;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.function.BiConsumer;
 import java.util.function.Function;
@@ -79,64 +72,12 @@ public class ForgeNetworkHandler implements NetworkHandler {
     }
 
     @Override
-    public void sendToServer(Message<?> message) {
-        Objects.requireNonNull(Minecraft.getInstance().getConnection(), "Cannot send packets when not in game!");
-        Minecraft.getInstance().getConnection().send(this.toServerboundPacket(message));
-    }
-
-    @Override
-    public void sendTo(Message<?> message, ServerPlayer player) {
-        player.connection.send(this.toClientboundPacket(message));
-    }
-
-    @Override
-    public void sendToAll(Message<?> message) {
-        Proxy.INSTANCE.getGameServer().getPlayerList().broadcastAll(this.toClientboundPacket(message));
-    }
-
-    @Override
-    public void sendToAllExcept(Message<?> message, ServerPlayer exclude) {
-        final Packet<?> packet = this.toClientboundPacket(message);
-        for (ServerPlayer player : Proxy.INSTANCE.getGameServer().getPlayerList().getPlayers()) {
-            if (player != exclude) {
-                player.connection.send(packet);
-            }
-        }
-    }
-
-    @Override
-    public void sendToAllNear(Message<?> message, BlockPos pos, Level level) {
-        this.sendToAllNearExcept(message, null, pos.getX(), pos.getY(), pos.getZ(), 64.0, level);
-    }
-
-    @Override
-    public void sendToAllNearExcept(Message<?> message, @Nullable ServerPlayer exclude, double posX, double posY, double posZ, double distance, Level level) {
-        Proxy.INSTANCE.getGameServer().getPlayerList().broadcast(exclude, posX, posY, posZ, distance, level.dimension(), this.toClientboundPacket(message));
-    }
-
-    @Override
-    public void sendToDimension(Message<?> message, Level level) {
-        this.sendToDimension(message, level.dimension());
-    }
-
-    @Override
-    public void sendToDimension(Message<?> message, ResourceKey<Level> dimension) {
-        Proxy.INSTANCE.getGameServer().getPlayerList().broadcastAll(this.toClientboundPacket(message), dimension);
-    }
-
-    /**
-     * @param message   message to create packet from
-     * @return          packet for message
-     */
-    private Packet<?> toServerboundPacket(Message<?> message) {
+    public Packet<?> toServerboundPacket(Message<?> message) {
         return this.channel.toVanillaPacket(message, NetworkDirection.PLAY_TO_SERVER);
     }
 
-    /**
-     * @param message   message to create packet from
-     * @return          packet for message
-     */
-    private Packet<?> toClientboundPacket(Message<?> message) {
+    @Override
+    public Packet<?> toClientboundPacket(Message<?> message) {
         return this.channel.toVanillaPacket(message, NetworkDirection.PLAY_TO_CLIENT);
     }